refactor Database and DatabaseHolder into /database
