+++ app/src/androidTest/java/com/github/fribourgsdp/radio/PlaylistSongsFragmentTest.kt
@@ -25,6 +25,7 @@ import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.data.view.AddPlaylistActivity
 import com.github.fribourgsdp.radio.data.view.PLAYLIST_DATA
 import com.github.fribourgsdp.radio.data.view.PlaylistSongsFragment
+import com.github.fribourgsdp.radio.database.Database
 import com.google.android.gms.tasks.Tasks
 import junit.framework.Assert.assertTrue
 import org.junit.Assert.assertEquals
