+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockLobbyActivities.kt
@@ -2,10 +2,10 @@ package com.github.fribourgsdp.radio.mockimplementations
 
 import android.content.Intent
 import android.os.Bundle
-import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.database.Database
 import com.github.fribourgsdp.radio.game.Game
 import com.github.fribourgsdp.radio.game.prep.*
 import kotlinx.serialization.encodeToString
