--- app/src/main/java/com/github/fribourgsdp/radio/DatabaseHolder.kt
@@ -1,7 +1,0 @@
-package com.github.fribourgsdp.radio
-
-interface DatabaseHolder {
-    fun initializeDatabase() : Database{
-        return FirestoreDatabase()
-    }
-}
