+++ app/src/main/java/com/github/fribourgsdp/radio/FirestoreDatabase.kt
@@ -6,6 +6,7 @@ import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.database.Database
 import com.github.fribourgsdp.radio.game.Game
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
