+++ app/src/main/java/com/github/fribourgsdp/radio/database/DatabaseHolder.kt
@@ -1,0 +1,9 @@
+package com.github.fribourgsdp.radio.database
+
+import com.github.fribourgsdp.radio.FirestoreDatabase
+
+interface DatabaseHolder {
+    fun initializeDatabase() : Database {
+        return FirestoreDatabase()
+    }
+}
