+++ app/src/main/java/com/github/fribourgsdp/radio/game/handler/GameHandler.kt
@@ -1,6 +1,6 @@
 package com.github.fribourgsdp.radio.game.handler
 
-import com.github.fribourgsdp.radio.Database
+import com.github.fribourgsdp.radio.database.Database
 import com.github.fribourgsdp.radio.game.GameView
 import com.google.firebase.firestore.DocumentSnapshot
 import com.google.firebase.firestore.EventListener
