+++ app/src/main/java/com/github/fribourgsdp/radio/game/handler/HostGameHandler.kt
@@ -2,6 +2,7 @@ package com.github.fribourgsdp.radio.game.handler
 
 import android.util.Log
 import com.github.fribourgsdp.radio.*
+import com.github.fribourgsdp.radio.database.Database
 import com.github.fribourgsdp.radio.game.Game
 import com.github.fribourgsdp.radio.game.GameView
 import com.google.firebase.firestore.DocumentSnapshot
